MAINT: fix compiler warnings for _swt.pyx and _dwt.pyx
They were of this form (with Clang):

    pywt/_extensions/_swt.c:5163:59: warning: comparison of integers of different
    signs: 'int' and 'size_t'
          (aka 'unsigned long') [-Wsign-compare]
            for (__pyx_t_13 = (__pyx_v_start_level + 1); __pyx_t_13 < __pyx_t_16;

+++ pywt/_extensions/_dwt.pyx
@@ -233,7 +233,8 @@ cpdef idwt_axis(np.ndarray coefs_a, np.ndarray coefs_d,
 
     return output
 
-cpdef upcoef(bint do_rec_a, data_t[::1] coeffs, Wavelet wavelet, int level, int take):
+cpdef upcoef(bint do_rec_a, data_t[::1] coeffs, Wavelet wavelet, int level,
+             size_t take):
     cdef data_t[::1] rec
     cdef int i, retval
     cdef size_t rec_len, left_bound, right_bound, coeffs_size
